refactor: util.ts -> getInside.ts로 변경
