+++ frontend/src/types/getInside.ts
